server/requirements.txt
django
djangorestframework
markdown
django-filter
django-rest-knox